move Detect class to enums.py
