+++ circleguard/__init__.py
@@ -1,5 +1,5 @@
 from circleguard.circleguard import Circleguard, set_options
 from circleguard.replay import Check, Replay, ReplayMap, ReplayPath
-from circleguard.detect import Detect
+from circleguard.enums import Detect
 
 __version__ = "2.0d"
