--- circleguard/detect.py
@@ -1,10 +1,0 @@
-from enum import Flag
-
-class Detect(Flag):
-                   # (in binary)
-    STEAL = 1 << 0 # 0001
-    RELAX = 1 << 1 # 0010
-    REMOD = 1 << 2 # 0100
-
-    ALL = STEAL | RELAX | REMOD
-    NONE = 0
